solana: rename path
