+++ solana/programs/matching-engine/src/events/auction_updated.rs
