+++ solana/programs/matching-engine/src/events/mod.rs
@@ -1,8 +1,8 @@
 mod auction_settled;
 pub use auction_settled::*;
 
-mod auction_update;
-pub use auction_update::*;
+mod auction_updated;
+pub use auction_updated::*;
 
 mod order_executed;
 pub use order_executed::*;
